Enable in-place aug for augment_images() child calls
Previously, all calls to augment_images() would perform
validation, normalization and copying of inputs.
This is now changed to only happen for the first call
of augment_images() in each hierarchy of calls.
All child calls augment (by default) in-place and skip
the validation and normalization steps. This improves
the performance by skipping unnecessary validations,
normalizations and copies.
Note that this only happens if hooks is not set,
otherwise the mentioned steps are still performed.